Go: Move `getImportPath` to shared `util` package
